move entity action base class to lib folder
