+++ src/backoffice/documents/documents/entity-actions/publish.action.ts
@@ -1,5 +1,5 @@
 import { UmbDocumentRepository } from '../repository/document.repository';
-import { UmbEntityActionBase } from '../../../shared/entity-actions';
+import { UmbEntityActionBase } from '../../../../../libs/entity-action';
 import { UmbControllerHostInterface } from '@umbraco-cms/controller';
 
 export class UmbPublishDocumentEntityAction extends UmbEntityActionBase<UmbDocumentRepository> {
