+++ src/backoffice/documents/documents/entity-actions/rollback.action.ts
@@ -1,5 +1,5 @@
 import { UmbDocumentRepository } from '../repository/document.repository';
-import { UmbEntityActionBase } from '../../../shared/entity-actions';
+import { UmbEntityActionBase } from '../../../../../libs/entity-action';
 import { UmbControllerHostInterface } from '@umbraco-cms/controller';
 
 export class UmbRollbackDocumentEntityAction extends UmbEntityActionBase<UmbDocumentRepository> {
