+++ src/backoffice/shared/entity-actions/copy/copy.action.ts
@@ -1,4 +1,4 @@
-import { UmbEntityActionBase } from '..';
+import { UmbEntityActionBase } from '../../../../../libs/entity-action';
 import { UmbControllerHostInterface } from '@umbraco-cms/controller';
 
 export class UmbCopyEntityAction<T extends { copy(): Promise<void> }> extends UmbEntityActionBase<T> {
