+++ src/backoffice/shared/entity-actions/move/move.action.ts
@@ -1,4 +1,4 @@
-import { UmbEntityActionBase } from '..';
+import { UmbEntityActionBase } from '../../../../../libs/entity-action';
 import { UmbControllerHostInterface } from '@umbraco-cms/controller';
 
 export class UmbMoveEntityAction<T extends { move(): Promise<void> }> extends UmbEntityActionBase<T> {
