+++ src/backoffice/shared/entity-actions/sort-children-of/sort-children-of.action.ts
@@ -1,4 +1,4 @@
-import { UmbEntityActionBase } from '..';
+import { UmbEntityActionBase } from '../../../../../libs/entity-action';
 import { UmbControllerHostInterface } from '@umbraco-cms/controller';
 
 export class UmbSortChildrenOfEntityAction<
